src/servo/dom/bindings/codegen/test/test_lookupGetter.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=462428
-->
<head>
  <title>Test for Bug 462428</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=462428">Mozilla Bug 462428</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 462428 **/
var x = new XMLHttpRequest;
x.open("GET", "");
var getter = x.__lookupGetter__('readyState');
ok(getter !== undefined, "But able to look it up the normal way");
ok(!x.hasOwnProperty('readyState'), "property should still be on the prototype");

var sawProp = false;
for (var i in x) {
    if (i === "readyState") {
        sawProp = true;
    }
}

ok(sawProp, "property should be enumerable");

is(getter.call(x), 1, "the getter actually works");

Object.getPrototypeOf(x).__defineSetter__('readyState', function() {});
is(getter.call(x), 1, "the getter works after defineSetter");

is(x.responseType, "", "Should have correct responseType up front");
var setter = x.__lookupSetter__('responseType');
setter.call(x, "document");
is(x.responseType, "document", "the setter is bound correctly");

</script>
</pre>
</body>
</html>
